done 4.22

+++ part4/bloglist/controllers/blogs.js
@@ -1,54 +1,86 @@
-const blogRouter = require('express').Router();
-const { BlogModel } = require('../models/blog');
+const blogsRouter = require('express').Router();
+const { Blog } = require('../models/blog');
 
 // retrieve all blogs
-blogRouter.get('/', async (request, response, next) => {
+blogsRouter.get('/', async (request, response, next) => {
   try {
-    const blogs = await BlogModel.find({});
-    response.json(blogs);
+    const blogs = await Blog.find({}).populate('user', {
+      username: 1,
+      name: 1,
+    });
+
+    if (!blogs) {
+      return response.status(404).json({ error: 'blogs not found' });
+    }
+
+    response.status(200).json(blogs);
   } catch (error) {
     next(error);
   }
 });
 
 // retrieve one blog
-blogRouter.get('/:id', async (request, response, next) => {
-  const id = request.params.id;
-
+blogsRouter.get('/:id', async (request, response, next) => {
   try {
-    const blog = await BlogModel.findById(id);
+    const blog = await Blog.findById(request.params.id).populate('user', {
+      username: 1,
+      name: 1,
+    });
 
     if (!blog) {
       return response.status(404).json({ error: 'blog not found' });
     }
 
-    response.json(blog);
+    response.status(200).json(blog);
   } catch (error) {
     next(error);
   }
 });
 
-// create blog
-blogRouter.post('/', async (request, response, next) => {
-  const { title, author, url, likes } = request.body;
+// create a blog
+blogsRouter.post('/', async (request, response, next) => {
+  const { title, url } = request.body;
+
+  const blogErrorMessage = {};
 
   if (!title) {
-    return response.status(400).json({ error: 'title is required' });
-  } else if (!author) {
-    return response.status(400).json({ error: 'author is required' });
-  } else if (!url) {
-    return response.status(400).json({ error: 'url is required' });
+    blogErrorMessage.title = 'title is required';
   }
 
-  const blog = new BlogModel({
-    title,
-    author,
-    url,
-    likes: likes || 0,
-  });
+  const urlRegex = /^(https?):\/\/[-\w\d.]+\.[A-Za-z]{2,}(\/.*)?$/;
+  if (!url) {
+    blogErrorMessage.url = 'url is required';
+  } else if (!urlRegex.test(url)) {
+    blogErrorMessage.url = 'url incorrect format';
+  }
+
+  if (Object.keys(blogErrorMessage).length > 0) {
+    return response.status(400).json({ error: blogErrorMessage });
+  }
 
   try {
+    const user = request.user;
+
+    if (!user) {
+      return response
+        .status(401)
+        .json({ error: 'invalid token - user not found' });
+    }
+
+    const blog = new Blog({
+      title,
+      author: user.name,
+      url,
+      likes: 0,
+      user: user._id,
+    });
+
     const savedBlog = await blog.save();
+
+    user.blogs = [...user.blogs, savedBlog._id];
+
+    await user.save();
+
     response.status(201).json(savedBlog);
   } catch (error) {
     next(error);
@@ -56,15 +88,32 @@ blogRouter.post('/', async (request, response, next) => {
 });
 
 // update blog
-blogRouter.put('/:id', async (request, response, next) => {
+blogsRouter.put('/:id', async (request, response, next) => {
   const { title, author, url } = request.body;
 
+  const blogError = {};
   if (!title) {
-    return response.status(400).json({ error: 'title is required' });
-  } else if (!author) {
-    return response.status(400).json({ error: 'author is required' });
-  } else if (!url) {
-    return response.status(400).json({ error: 'url is required' });
+    blogError.title = 'title is required';
+  } else if (title.trim().length < 3) {
+    blogError.title = 'title must be at least 3 characters long';
+  }
+
+  if (!author) {
+    blogError.author = 'author is required';
+  } else if (author.trim().length < 3) {
+    blogError.author = 'author must be at least 3 characters long';
+  }
+
+  const urlRegex =
+    /^(http|https):\/\/[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}(?:\/[^\s]*)?$/;
+  if (!url) {
+    blogError.author = 'url is required';
+  } else if (urlRegex.test(url)) {
+    blogError.url = 'invalid url format';
+  }
+
+  if (Object.keys(blogError).length > 0) {
+    return response.status(400).json({ error: blogError });
   }
 
   const blog = {
@@ -75,7 +124,7 @@ blogRouter.put('/:id', async (request, response, next) => {
 
   const id = request.params.id;
   try {
-    const updatedBlog = await BlogModel.findById(id);
+    const updatedBlog = await Blog.findById(id);
 
     if (!updatedBlog) {
       return response.status(404).json({ error: 'blog not found' });
@@ -83,28 +132,59 @@ blogRouter.put('/:id', async (request, response, next) => {
 
     updatedBlog.set(blog);
     const savedBlog = await updatedBlog.save();
-    response.json(savedBlog);
+    response.status(201).json(savedBlog);
   } catch (error) {
     next(error);
   }
 });
 
 // delete blog
-blogRouter.delete('/:id', async (request, response, next) => {
-  const id = request.params.id;
-
+blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    const blog = await BlogModel.findById(id);
+    const blog = await Blog.findById(request.params.id);
 
     if (!blog) {
       return response.status(404).json({ error: 'blog not found' });
     }
 
+    const user = request.user;
+
+    if (!user) {
+      return response
+        .status(401)
+        .json({ error: 'invalid token - user not found' });
+    }
+
+    if (blog.user.toString() !== user._id.toString()) {
+      return response
+        .status(403)
+        .json({ error: 'unauthorized to delete this blog' });
+    }
+
     await blog.deleteOne();
+
+    user.blogs = (user.blogs || []).filter(
+      (b) => b.toString() !== blog._id.toString()
+    );
+    await user.save();
+  } catch (error) {
+    next(error);
+  }
+});
+
+// ! debugging (don't include in testing)
+blogsRouter.delete('/', async (request, response, next) => {
+  try {
+    const blogs = await Blog.deleteMany({});
+
+    if (!blogs) {
+      return response.status(400).json({ error: 'no blogs found' });
+    }
+
     response.status(204).end();
   } catch (error) {
     next(error);
   }
 });
 
-module.exports = blogRouter;
+module.exports = blogsRouter;
